Move register and field definitions to tmc4671_regs.py
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
